requirements.txt
gradio
langchain
openai
faiss-cpu
python-docx
PyPDF2
tiktoken
gunicorn